4/18 주석 추가

+++ src/main/java/com/gitlog/service/PostService.java
@@ -1,6 +1,6 @@
 package com.gitlog.service;
 
-import com.amazonaws.services.s3.transfer.Upload;
+import com.amazonaws.services.s3.AmazonS3Client;
 import com.gitlog.config.UserDetailsImpl;
 import com.gitlog.config.uploader.Uploader;
 import com.gitlog.dto.AccountResponseDto;
@@ -8,7 +8,6 @@ import com.gitlog.dto.CommentResponseDto;
 import com.gitlog.dto.PostRequestDto;
 import com.gitlog.dto.PostResponseDto;
 import com.gitlog.model.Account;
-import com.gitlog.model.Comment;
 import com.gitlog.model.Post;
 import com.gitlog.repository.AccountRepository;
 import com.gitlog.repository.CommentRepository;
@@ -21,11 +20,12 @@ import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.multipart.MultipartFile;
 
 import javax.transaction.Transactional;
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
@@ -36,11 +36,14 @@ public class PostService {
     private final CommentRepository commentRepository;
     private final HeartRepository heartRepository;
     private final Uploader uploader;
+    private final AmazonS3Client amazonS3Client;
 
     //게시글 가져오기
     public Page<PostResponseDto> getPost(int page, int size, UserDetailsImpl userDetails){
         PageRequest pageRequest = PageRequest.of(page -1, size , Sort.by(Sort.Direction.DESC, "createdAt"));
+        // 페이지로 포스트를 가져 오고
         Page<Post> posts = postRepository.findAll(pageRequest);
+        // 바로 내려주는것이 아니라 responseDTO 에 넣어서 필요한 정보만 넣어서 Client 에게 넘겨줄 수 있도록
         Page<PostResponseDto> toMap = posts.map(post -> new PostResponseDto(
                 post.getId(),
                 post.getContent(),
@@ -119,15 +122,50 @@ public class PostService {
     }
     //게시글 삭제
     @Transactional
-    public ResponseEntity<String> deletePost(Long post_id){
+    public ResponseEntity<String> deletePost(Long post_id, Account account){
         Post post = postRepository.findById(post_id).orElse(null);
         if (post == null){
             return new ResponseEntity<>("해당 게시글은 없는 게시글입니다.", HttpStatus.BAD_REQUEST);
         }
-        postRepository.deleteById(post_id);
+        if (!post.getAccount().getNickname().equals(account.getNickname())){
+            return new ResponseEntity<>("다른 사용자의 게시글을 삭제 할 수 없습니다.", HttpStatus.BAD_REQUEST);
+        }
+        amazonS3Client.deleteObject("hanghae99-gitlog",post.getImgUrl());
+        post.deletePost(post);
         commentRepository.deleteAllByPost(post);
         heartRepository.deleteAllByPost(post);
-
+        postRepository.deleteById(post_id);
         return new ResponseEntity<>("성공적으로 삭제하였습니다", HttpStatus.OK);
+    }
+
+    public ResponseEntity userStory(String nickname){
+        Account account = accountRepository.findByNickname(nickname).orElse(null);
+        if (account == null){
+            return new ResponseEntity<>("없는 사용자 입니다", HttpStatus.BAD_REQUEST);
+        }
+        List<Post> posts = postRepository.findAllByAccount_Nickname(nickname);
+        List<PostResponseDto> toList = posts.stream().map(
+                post -> new PostResponseDto(
+                post.getId(),
+                post.getContent(),
+                post.getImgUrl(),
+                post.getCreatedAt(),
+                post.getCreatedBy(),
+                post.getModifiedAt(),
+                post.getComments().stream().map(
+                        comment -> new CommentResponseDto(
+                                comment.getId(),
+                                comment.getContent(),
+                                comment.getCreatedAt(),
+                                comment.getCreatedBy()
+                        )
+                ).collect(Collectors.toList()),
+                new AccountResponseDto(post.getAccount().getProfileImgUrl()),
+                post.getComments().size(),
+                post.getHearts().size()
+        )).collect(Collectors.toList());
+        Map<String, Object> map = new HashMap<>();
+        map.put("posts",toList);
+        return ResponseEntity.ok().body(map);
     }
 }
